tentative d'amélioration du timeout

+++ src/includes/stack_type.h
@@ -5,7 +5,7 @@
 struct model_elem {
   int element ;
   int duplicate;
-  double RTT;
+  int RTT;
   struct model_elem* next;
 };
